requirements.txt
undetected-chromedriver
setuptools
livekit
livekit-api
numpy
matplotlib
click
python-dotenv
pyyaml

